requirements.txt
streamlit
python-dotenv

# LangChain stack
langchain>=0.2
langchain-core>=0.2
langchain-openai
langchain-chroma==0.1.1

# Chroma (DuckDB backend, no SQLite requirement)
chromadb==0.4.24

# Misc
pydantic>=2
lxml